Add op parameter to DistributedOptimizers
Also change allreduce_type to op in allreduce functions. With these
changes using adasum is as easy as:
>hvd.allreduce(tensor, op=hvd.Adasum)
or
>opt = DistributedOptimizer(opt, op=hvd.Adasum)

No longer check for HOROVOD_ADASUM env var in framework code. The intent
is that HOROVOD_ADASUM will only be used to select algorithm
implementations, but we will default to something reasonable without it.
Setting that default is still todo.

Update docstrings to reflect new parameters.