Add process binding arguments to horovodrun
To maintain compute/memory locality, it is often good practise to
bind processes to NUMA nodes.
In ppc64le architecture with large models, I see a 20% performance gain
because tensor swapping happens between GPU memory and system memory and
CPU/GPU link (NVLINK) is faster than inter-socket link.
The PR also makes socket binding the default for SMPI (ppc64le). Even
with those flags, you may end up in a situation when a process with a GPU is bound to a
non-local CPU socket. That happens usually when GPUs are not split evenly across sockets or
you don't use all the GPUs of a node. In those cases, it
is recommended to use a rankfile to get more control: --binding-args="--rankfile myrankfile"

Signed-off-by: Nicolas V Castet <nvcastet@us.ibm.com>