Fix for ROCm breakage - 200805
The following commit breaks ROCm support for horovod-tensorflow

https://github.com/horovod/horovod/commit/7a1c761c663eb758536c1bff62d9d9c9a032dd38

horovod build fails because the `<cuda_runtime.h>` cannot be found on the ROCm platform

This PR/commit fixes the above breakage by including the HIP runtime file (instead of the CUDA one), when compiling with ROCm.

In addition to the above change, this commit also renames a couple of cuda* names to the corresponding gpu* versions instead.

Signed-off-by: Deven Desai <deven.desai.amd@gmail.com>