Fix hvd.barrier() tensor queue management
If a barrier was enqueued together with an allreduce, the barrier request
would erroneously be pushed again and again into the tensor queue each
cycle even after the barrier was completed. Additionally, this could
cause a mismatch in GetTensorEntriesFromResponse where a tensor from
the response would not be present in the tensor table. This caused
segmentation faults in the test_horovod_join_allgather unit test
although the situation was established earlier by a hvd.barrier() in the
test_horovod_allreduce_duplicate_name_error unit test.

Signed-off-by: Max H. Gerlach <git@maxgerlach.de>